renamed lengthy class names
