+++ jobs/v4/src/main/java/com/example/jobs/CommuteSearchJobs.java
@@ -33,7 +33,7 @@ import com.google.type.LatLng;
 
 import java.io.IOException;
 
-public class JobSearchCommuteSearch {
+public class CommuteSearchJobs {
 
   public static void searchJobs() throws IOException {
     // TODO(developer): Replace these variables before running the sample.
