+++ jobs/v4/src/main/java/com/example/jobs/HistogramSearchJobs.java
@@ -29,7 +29,7 @@ import com.google.cloud.talent.v4beta1.TenantOrProjectName;
 
 import java.io.IOException;
 
-public class JobSearchHistogramSearch {
+public class HistogramSearchJobs {
 
   public static void searchJobs() throws IOException {
     // TODO(developer): Replace these variables before running the sample.
